udi-87 PR changes. Added abstract event.
