+++ model/event/DacAddedEvent.php
@@ -15,7 +15,7 @@
  * along with this program; if not, write to the Free Software
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  *
- * Copyright (c) 2016 (original work) Open Assessment Technologies SA
+ * Copyright (c) 2020 (original work) Open Assessment Technologies SA
  *
  */
 
@@ -24,75 +24,6 @@ namespace oat\taoDacSimple\model\event;
 use JsonSerializable;
 use oat\oatbox\event\Event;
 
-class DacAddedEvent implements Event, JsonSerializable
+class DacAddedEvent extends AbstractDacEvent implements Event, JsonSerializable
 {
-
-    /** @var  string */
-    protected $userUri;
-    /** @var string */
-    protected $resourceUri;
-    /** @var string[]  */
-    protected $privilege;
-
-    /**
-     * @param string   $userUri
-     * @param string   $resourceUri
-     * @param string[] $privilege
-     */
-    public function __construct(string $userUri, string $resourceUri, array $privilege)
-    {
-        $this->userUri = $userUri;
-        $this->resourceUri = $resourceUri;
-        $this->privilege = $privilege;
-    }
-
-
-    /**
-     * Return a unique name for this event
-     * @see \oat\oatbox\event\Event::getName()
-     */
-    public function getName()
-    {
-        return get_class($this);
-    }
-
-    /**
-     * Specify data which should be serialized to JSON
-     * @link  http://php.net/manual/en/jsonserializable.jsonserialize.php
-     * @return mixed data which can be serialized by <b>json_encode</b>,
-     * which is a value of any type other than a resource.
-     * @since 5.4.0
-     */
-    public function jsonSerialize()
-    {
-        return [
-            'userUri'   => $this->userUri,
-            'accessUri' => $this->resourceUri,
-            'privilege' => $this->privilege
-        ];
-    }
-
-    /**
-     * @return string
-     */
-    public function getUserUri(): string
-    {
-        return $this->userUri;
-    }
-
-    /**
-     * @return string
-     */
-    public function getResourceUri(): string
-    {
-        return $this->resourceUri;
-    }
-
-    /**
-     * @return string[]
-     */
-    public function getPrivilege(): array
-    {
-        return $this->privilege;
-    }
 }
