+++ model/event/DacRootAddedEvent.php
@@ -28,75 +28,6 @@ use oat\oatbox\event\Event;
  * Class DacRootAddedEvent This event not triggered on adding child permissions with recursive option
  * @package oat\taoDacSimple\model\event
  */
-class DacRootAddedEvent implements Event, JsonSerializable
+class DacRootAddedEvent extends AbstractDacEvent implements Event, JsonSerializable
 {
-
-    /** @var  string */
-    protected $userUri;
-    /** @var string */
-    protected $resourceUri;
-    /** @var string[]  */
-    protected $privilege;
-
-    /**
-     * @param string   $userUri
-     * @param string   $resourceUri
-     * @param string[] $privilege
-     */
-    public function __construct(string $userUri, string $resourceUri, array $privilege)
-    {
-        $this->userUri = $userUri;
-        $this->resourceUri = $resourceUri;
-        $this->privilege = $privilege;
-    }
-
-
-    /**
-     * Return a unique name for this event
-     * @see \oat\oatbox\event\Event::getName()
-     */
-    public function getName()
-    {
-        return get_class($this);
-    }
-
-    /**
-     * Specify data which should be serialized to JSON
-     * @link  http://php.net/manual/en/jsonserializable.jsonserialize.php
-     * @return mixed data which can be serialized by <b>json_encode</b>,
-     * which is a value of any type other than a resource.
-     * @since 5.4.0
-     */
-    public function jsonSerialize()
-    {
-        return [
-            'userUri'   => $this->userUri,
-            'accessUri' => $this->resourceUri,
-            'privilege' => $this->privilege
-        ];
-    }
-
-    /**
-     * @return string
-     */
-    public function getUserUri(): string
-    {
-        return $this->userUri;
-    }
-
-    /**
-     * @return string
-     */
-    public function getResourceUri(): string
-    {
-        return $this->resourceUri;
-    }
-
-    /**
-     * @return string[]
-     */
-    public function getPrivilege(): array
-    {
-        return $this->privilege;
-    }
 }
